+++ tortoise_cli/cli.py
@@ -1,35 +1,23 @@
 from __future__ import annotations
 
-import asyncio
 import contextlib
-import os
 import sys
-from functools import wraps
-from pathlib import Path
+from collections.abc import AsyncGenerator
 
-import click
+import asyncclick as click
 from ptpython.repl import embed
-from tortoise import Tortoise
+from tortoise import Tortoise, connections
 
 from tortoise_cli import __version__, utils
 
-if sys.version_info >= (3, 11):
-    import tomllib
-else:
-    import tomlkit as tomllib
 
-
-def coro(f):
-    @wraps(f)
-    def wrapper(*args, **kwargs):
-        loop = asyncio.get_event_loop()
-        try:
-            loop.run_until_complete(f(*args, **kwargs))
-        finally:
-            if f.__name__ != "cli":
-                loop.run_until_complete(Tortoise.close_connections())
-
-    return wrapper
+@contextlib.asynccontextmanager
+async def aclose_tortoise() -> AsyncGenerator[None]:
+    try:
+        yield
+    finally:
+        if Tortoise._inited:
+            await connections.close_all()
 
 
 @click.group(context_settings={"help_option_names": ["-h", "--help"]})
@@ -40,16 +28,8 @@ def coro(f):
     help="TortoiseORM config dictionary path, like settings.TORTOISE_ORM",
 )
 @click.pass_context
-@coro
-async def cli(ctx: click.Context, config: str | None):
-    if (
-        not config
-        and not (config := os.getenv("TORTOISE_ORM"))
-        and (p := Path("pyproject.toml")).exists()
-    ):
-        doc = tomllib.loads(p.read_text("utf-8"))
-        config = doc.get("tool", {}).get("aerich", {}).get("tortoise_orm", "")
-    if not config:
+async def cli(ctx: click.Context, config: str | None) -> None:
+    if not config and not (config := utils.tortoise_orm_config()):
         raise click.UsageError(
             "You must specify TORTOISE_ORM in option or env, or config file pyproject.toml from config of aerich",
             ctx=ctx,
@@ -61,19 +41,19 @@ async def cli(ctx: click.Context, config: str | None):
 
 @cli.command(help="Start a interactive shell.")
 @click.pass_context
-@coro
-async def shell(ctx: click.Context):
-    with contextlib.suppress(EOFError, ValueError):
-        await embed(
-            globals=globals(),
-            title="Tortoise Shell",
-            vi_mode=True,
-            return_asyncio_coroutine=True,
-            patch_stdout=True,
-        )
+async def shell(ctx: click.Context) -> None:
+    async with aclose_tortoise():
+        with contextlib.suppress(EOFError, ValueError):
+            await embed(
+                globals=globals(),
+                title="Tortoise Shell",
+                vi_mode=True,
+                return_asyncio_coroutine=True,
+                patch_stdout=True,
+            )
 
 
-def main():
+def main() -> None:
     if sys.path[0] != ".":
         sys.path.insert(0, ".")
     cli()